Move the cursor to its own file
